+++ src/main/integrations/electron-minidump.ts
@@ -7,7 +7,7 @@ import { app, crashReporter } from 'electron';
 import { mergeEvents, normalizeEvent } from '../../common';
 import { getEventDefaults } from '../context';
 import { onRendererProcessGone, rendererRequiresCrashReporterStart, usesCrashpad } from '../electron-normalize';
-import { sessionCrashed } from './main-process-session';
+import { checkPreviousSession, sessionCrashed, unreportedDuringLastSession } from '../sessions';
 
 /** Is object defined and has keys */
 function hasKeys(obj: any): boolean {
@@ -82,13 +82,18 @@ export class ElectronMinidump implements Integration {
 
     // If a renderer process crashes, mark any existing session as crashed
     onRendererProcessGone((_, __) => {
-      sessionCrashed({ forceCapture: true });
+      sessionCrashed();
     });
 
     // If we're using the Crashpad minidump uploader, we set extra parameters whenever the scope updates
     if (usesCrashpad()) {
       this._setupScopeListener();
     }
+
+    // Check if last crash report was likely to have been unreported in the last session
+    const previousSessionCrashed = unreportedDuringLastSession(crashReporter.getLastCrashReport()?.date);
+    // Check if a previous session was not closed
+    forget(checkPreviousSession(previousSessionCrashed));
   }
 
   /**