+++ src/main/integrations/main-process-session.ts
@@ -1,10 +1,8 @@
-import { getCurrentHub } from '@sentry/core';
-import { flush } from '@sentry/node';
-import { Integration, SessionStatus } from '@sentry/types';
+import { Integration } from '@sentry/types';
 import { logger } from '@sentry/utils';
 import { app } from 'electron';
 
-const TERMINAL_STATES = [SessionStatus.Exited, SessionStatus.Crashed];
+import { endSession, startSession } from '../sessions';
 
 /** Tracks sessions as the main process lifetime. */
 export class MainProcessSession implements Integration {
@@ -16,9 +14,7 @@ export class MainProcessSession implements Integration {
 
   /** @inheritDoc */
   public setupOnce(): void {
-    const hub = getCurrentHub();
-    logger.log('MainProcessSession - Start session');
-    hub.startSession();
+    startSession();
 
     // We track sessions via the 'will-quit' event which is the last event emitted before close.
     //
@@ -43,41 +39,15 @@ export class MainProcessSession implements Integration {
 
   /** Handles the exit */
   private _exitHandler: (event: Electron.Event) => Promise<void> = async (event: Electron.Event) => {
-    logger.log('MainProcessSession - Exit Handler');
+    logger.log('[MainProcessSession] Exit Handler');
 
     // Stop the exit so we have time to send the session
     event.preventDefault();
-    const hub = getCurrentHub();
 
-    const session = hub.getScope()?.getSession();
-
-    if (session && !TERMINAL_STATES.includes(session.status)) {
-      logger.log('MainProcessSession - Ending session');
-      hub.endSession();
-    } else {
-      logger.log('MainProcessSession - Session was already ended', session);
-    }
-
-    await flush();
+    // End the session
+    await endSession();
 
     // After flush we can safely exit
     app.exit();
   };
-}
-
-/** Sets the current session as crashed */
-export function sessionCrashed(options: { forceCapture?: boolean } = {}): void {
-  logger.log('Session Crashed');
-  const hub = getCurrentHub();
-  const session = hub.getScope()?.getSession();
-
-  if (session && !TERMINAL_STATES.includes(session.status)) {
-    session.update({ status: SessionStatus.Crashed, errors: (session.errors += 1) });
-  } else {
-    logger.log('No session to update');
-  }
-
-  if (options.forceCapture) {
-    hub.captureSession();
-  }
 }