+++ src/main/integrations/sentry-minidump/index.ts
@@ -12,9 +12,9 @@ import { ElectronNetTransport } from '../../transports/electron-net';
 import { BaseUploader } from './base-uploader';
 import { BreakpadUploader } from './breakpad-uploader';
 import { CrashpadUploader } from './crashpad-uploader';
-import { Store } from './store';
+import { Store } from '../../store';
 import { getEventDefaults } from '../../context';
-import { sessionCrashed } from '../main-process-session';
+import { checkPreviousSession, sessionCrashed } from '../../sessions';
 
 /** Sends minidumps via the Sentry uploader.. */
 export class SentryMinidump implements Integration {
@@ -80,7 +80,11 @@ export class SentryMinidump implements Integration {
         level: Severity.Fatal,
         platform: 'native',
         tags: { 'event.environment': 'native', 'event.process': 'browser', event_type: 'native' },
-      }),
+      }).then((minidumpsFound) =>
+        // Check for previous uncompleted session. If a previous session exists
+        // and no minidumps were found, its likely an abnormal exit
+        checkPreviousSession(minidumpsFound),
+      ),
     );
   }
 
@@ -131,8 +135,8 @@ export class SentryMinidump implements Integration {
       tags: { 'event.environment': 'native', 'event.process': crashedProcess, event_type: 'native' },
     });
 
-    await sessionCrashed();
     await this._sendNativeCrashes(options, event);
+    sessionCrashed();
 
     addBreadcrumb({
       category: 'exception',
@@ -159,8 +163,12 @@ export class SentryMinidump implements Integration {
     }
   }
 
-  /** Loads new native crashes from disk and sends them to Sentry. */
-  private async _sendNativeCrashes(options: ElectronMainOptions, event: Event): Promise<void> {
+  /**
+   * Loads new native crashes from disk and sends them to Sentry.
+   *
+   * Returns true if one or more minidumps were found
+   */
+  private async _sendNativeCrashes(options: ElectronMainOptions, event: Event): Promise<boolean> {
     // Whenever we are called, assume that the crashes we are going to load down
     // below have occurred recently. This means, we can use the same event data
     // for all minidumps that we load now. There are two conditions:
@@ -195,7 +203,7 @@ export class SentryMinidump implements Integration {
             logger.warn(
               `Discarding event because it's not included in the random sample (sampling rate = ${sampleRate})`,
             );
-            return;
+            return true;
           }
 
           if (newEvent && beforeSend) {
@@ -203,7 +211,7 @@ export class SentryMinidump implements Integration {
 
             if (beforeSendResult === null) {
               logger.warn('`beforeSend` returned `null`, will not send event.');
-              return;
+              return true;
             }
 
             newEvent = beforeSendResult;
@@ -215,10 +223,13 @@ export class SentryMinidump implements Integration {
         }
         // Unset to recover memory
         this._scopeLastRun = undefined;
+        return true;
       }
     } catch (_oO) {
       logger.error('Error while sending native crash.');
     }
+
+    return false;
   }
 
   /**